requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
fuzzywuzzy==0.18.0
idna==3.10
Levenshtein==0.26.1
numpy==2.1.3
pandas==2.2.3
PyMuPDF==1.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
six==1.17.0
tqdm==4.67.1
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
